w3c-reports/12092014/gtk/runned-tests/websockets/Create-nonAscii-protocol-string-expected.txt

CONSOLE MESSAGE: Wrong protocol for WebSocket '\u0080echo'

PASS W3C WebSocket API - Create WebSocket - Pass a valid URL and a protocol string with non-ascii values - SYNTAX_ERR is thrown 
